Clean and consolidation
